requirements.txt
Pillow==10.4.0
screeninfo==0.8.1